Implements DNSStatus tile functionality
Now the DNS status tile shows basic information about the Blocky API.
A basic query is executed to the endpoint API to check if the API is
properly working and if the name resolution from the DNS server works
properly.
Furthermore, the UDP (DNS) and TCP (API) port states are probed for extra
information.